move duplicated code into central place
